Allow '-' to be passed as input and output.
Separate input and output and allow '-' symbols to be passed as
the names of files specifying that stdin and stdout should be used
for input and output respecitvely.

Also move all debug/verbose and other service output from stdout
to stderr to allow the complete solution be passed to stdout.

This patch allows minisat to be incorporated with other unix utilites
that uses pipes for communication and not files.